project/requirements.txt
llama-index-core>=0.10.0
llama-index-llms-anthropic>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-parse>=0.4.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.18.0
sentence-transformers>=2.2.0
torch>=2.0.0
ipython>=8.0.0
requests>=2.31.0
PyPDF2>=3.0.1
PyYAML>=6.0.0
pytest>=8.4.2

